Add Whitespace fragments.
We now have explicit support for Fragments.
Fragments should be parsed with a FragmentParser.
For now we only have a whitespace fragment parser, which uses the old whitespace hack to get the offsets right.
But we can extend the set of parsers in the future.

Also cleaned up the places where error happen.
Next to errors from parsing and analyzing using the SPT language,
the TestCaseExtractor now also checks simple semantics not related to executing the test.
For now the only check is to ensure that all expectations actually have an ITestExpectation that can evaluate it.

The goal of this is to have all errors relating to the SPT side of the test case in the TestCaseExtractor.
I want the TestRunner to only return messages related to actually running the test case.
Like messages of failed expectations, or errors relating to the language under test with which the test is being run.