Add `resolve` and `run` expectations.
As both `parse to` and `run strat to` used the same part after `to`,
it is now extracted in a FragmentUtil.
This class can be injected and used by test expectations to parse or analyze fragments inside the `to` part and generate error messages if things go wrong.
For now it's used by `parse to` and `run to`, but if we add `analyze to` and implement `build builder to`,
they too will be able to use this utility class.

Analysis of these fragments required access to a Project, so that was added to tests and fragments.
The resource of a test and fragment became mandatory (i.e., not null).
They already were, as analysis requires a resource, but they were still marked and treated as Nullable.