Add simplistic ATerm support to SPT.
Had to add a new ITestExpectation for each 'x to ATerm' expectation, as it is Spoofax specific, and therefore not allowed to be part of the generic part of SPT. These expectation evaluators are very similar to the evaluator for the 'x to fragment' expecations. There is a lot of code duplication, but for now I'm ok with that.