Split SPT into MetaBorg Testing and SPoofax Testing.
MBT is the part that only depends on Metaborg Core and is a framework for the implementation of Language Parametric Testing Languages (LPTL) like SPT.

MBT provides an implementation independent (i.e. not Spoofax specific) model for tests, and interfaces for the extraction from some language to this model, and for the running of these tests.
It also provides a set of generic expectations that should (probably) be part of any language implementation.

SPT-core implements these interfaces and allows extraction of test cases from an SPT specification, and running tests on Spoofax languages.
It also supplies some Spoofax specific expectations (like the running of Stratego strategies on fragments, or the comparison against ATerms).