Add a FragmentParserConfig to test execution.
This config can influence how the fragments in the test case should be parsed.
At the moment, we have a spoofax specific config which can hold a JSGLRParserConfigurator for a language.
All fragments of that language will get parsed with that config.
This allows us to change the start symbol for the language under test for the tests in a suite.