Relax selection constraints.
`run on #i` expectations will now be tried on each outer node within the selection `i`.
It used to be an exact match by offset, but this does not allow selecting any node that starts with `[` or ends with `]`.

Also updated the `resolve #i to #j` to perform resolution in the middle of selection `i`, instead of at the start.
I hope that will allow the same kind of 'wide selections'.