Fix issue with reference resolution tests.
The `resolve to` expectation checked to make sure that it resolved to the correct file, using `FileObject.equals`.
The first problem was that this method is the same as using `==`, so it's an identity check, not a check for the same file.
The second problem was that the error message was not detailed enough to report what actually went wrong.