Fix NPE in parse-to expectation.
When evaluating a parse-to expectation where the input fragment failed to parse and the resulting AST was null, it would still be compared to the output fragment, resulting in an NPE.

Due to error recovery of Spoofax, usually the AST is not null, even if parsing failed, so this was tricky to discover. Yay for Compiler Construction labs.