Update expectation providers.
The test case extraction phase has to 'parse' the `IStrategoTerms`
into SPT model objects.
This creates a coupling between the SPT grammar and the test extractor.
I forgot to update some of the test expectation providers
when I changed the grammar to add the `on #n` part to run expectations.
This commit adds a test case for regression testing,
and updates the expectation providers to check more thoroughly if they can handle the AST.