Update SPT testrunner for Eclipse.
The Eclipse testrunner should now be working properly.
It still misses cool stack traces and the 'show only failed tests' filter is broken,
but it's a good start.
To use it, right click on a project with tests, then in the context menu `Spoofax -> Run all SPT tests`.
This invocation procedure is up for debate.
All feedback is very welcome.