Fix test extraction.
The previous fix for test extraction introduced a bug in the extraction of analysis message expectations.

I also used this chance to add an `analysis succeeds` and `analysis fails` expectation.
These are, respectively, equivalent to `0 errors` and `> 0 errors`.
They have been added, because the syntactic sugar is pretty nice and consistent with the `parse` expectations.