Update the SPT testrunner in Eclipse.
It can now be invoked 3 different ways:
1. through the project context menu (right click) in the `Spoofax (meta)` submenu.
   In this case it runs all SPT files under the project you right clicked.
2. through the `Spoofax (meta)` menu bar entry, then select `Run tests of SPT file`.
   In this case it will check if you currently have an SPT file open, and if so, run it.
3. through the `Spoofax (meta)` menu bar entry, then select `Run all selected tests`.
   In this case it will check what you selected.
   If you selected any SPT files, directories, or projects, it will scan all of the selections and run all of the SPT files it found within those selections.

The testrunner UI has been updated.
Here's a quick documentation on how it works.

It will start displaying any test suites and test cases within those test suites as soon as it discovers them.
Then, after they are all loaded, they will be run one by one, and the progress bar at the top will increase.
As long as the progress bar remains green, no tests failed yet.
As soon as a test fails, the progress bar will turn red to indicate so.

The numbers next to the progress bar also indicate the progress.
For example, `5 / 7` means 5 tests passed already out of a total of 7 tests.
This can mean that either 2 tests failed, or some tests have not been executed yet.
Which case applies can be determined by looking at the progress bar.

Any selections that don't contain any SPT files are ignored.
Any SPT files that fail to extract will be included in the list on the bottom left side, along with the test suites that did manage to get extracted.
The ones that did not extract properly will be displayed in red, as opposed to the default black color for test suites.
By selecting a red test suite, the extraction errors will be displayed in the console on the bottom right.
Any test suite can be double clicked to open the corresponding file in Eclipse.
Test suites that got extracted succesfully can be expanded if they contained any test cases.
This will show all the test cases of that suite as child elements of the test suite in the bottom right view.

Test cases are displayed in the default black color if they have not been executed yet.
Test cases that have finished will have their duration appended to their name.
Failed test cases are displayed in red, and passing test cases are displayed in green.
A red test case can be selected, doing so will show the messages about the test failure,
including the exceptions that caused them (e.g. a `StrategoRuntimeException` with stacktrace) in the console on the bottom right.
Double clicking a test case will open the corresponding SPT file and jump to the location of the test case.

When a test case fails, the test suite that contained the failing test case will be appended with the number of failed tests in that test suite so far.